integration: move batch prepare tests to batch.rs
